extract tests to separate file
